Seguimiento_04_23 (#16)
* Interfaces graficas y conexiones iniciadas

* Commit desde David con Perfil_cliente DAO

* Interfaces gráficas de Inicio y Menu_principal listas (sin funcionalidades) (#2)

* Fer (#3)

* Creacion de la clase Ejercicios con los arrays de los ejercicios

* Introduccion de metodos a realizar para completar la clase dieta

* David (#4)

* Interfaces gráficas de Inicio y Menu_principal listas (sin funcionalidades)

* Crea perfil, inicia sesión y llega al menu principal con info del cliente.

* Seguimiento_02_23 (#5)

* Interfaces gráficas de Inicio y Menu_principal listas (sin funcionalidades)

* Crea perfil, inicia sesión y llega al menu principal con info del cliente.

* Funcionalidad de Editar_perfil.java añadida

* Funcionalidad de Editar_perfil.java añadida

* Archivo SQL añadido (#8)

* Interfaces gráficas de Inicio y Menu_principal listas (sin funcionalidades)

* Crea perfil, inicia sesión y llega al menu principal con info del cliente.

* Funcionalidad de Editar_perfil.java añadida

* Funcionalidad de Editar_perfil.java añadida

* Esquema SQL y datos para la base de datos añadidos

* Update proyecto_clientes.sql

* David (#9)

* Interfaces gráficas de Inicio y Menu_principal listas (sin funcionalidades)

* Crea perfil, inicia sesión y llega al menu principal con info del cliente.

* Funcionalidad de Editar_perfil.java añadida

* Funcionalidad de Editar_perfil.java añadida

* Esquema SQL y datos para la base de datos añadidos

* Dump

* Dump 2.0

* Username verificación OK

* David (#10)

* Interfaces gráficas de Inicio y Menu_principal listas (sin funcionalidades)

* Crea perfil, inicia sesión y llega al menu principal con info del cliente.

* Funcionalidad de Editar_perfil.java añadida

* Funcionalidad de Editar_perfil.java añadida

* Esquema SQL y datos para la base de datos añadidos

* Dump

* Dump 2.0

* Username verificación OK

* Cuadro Dieta OK

* David (#13)

* Interfaces gráficas de Inicio y Menu_principal listas (sin funcionalidades)

* Crea perfil, inicia sesión y llega al menu principal con info del cliente.

* Funcionalidad de Editar_perfil.java añadida

* Funcionalidad de Editar_perfil.java añadida

* Esquema SQL y datos para la base de datos añadidos

* Dump

* Dump 2.0

* Username verificación OK

* Cuadro Dieta OK

* Cuadro de diálogo y atributo sexo añadidos

* Fer (#14)

* Creacion de la clase Ejercicios con los arrays de los ejercicios

* Introduccion de metodos a realizar para completar la clase dieta

* resolucion de metodos y FORM de la dieta

* introduccion de los arrays y  metodos de ejercicios y comidas para seleccionarse de forma aleatoria

* arreglo de problemas en metodos de dieta

* commit

* forms y codigo de dieta

* Add files via upload

* Seguimiento_04_23 (#15)

* Interfaces gráficas de Inicio y Menu_principal listas (sin funcionalidades)

* Crea perfil, inicia sesión y llega al menu principal con info del cliente.

* Funcionalidad de Editar_perfil.java añadida

* Funcionalidad de Editar_perfil.java añadida

* Esquema SQL y datos para la base de datos añadidos

* Dump

* Dump 2.0

* Username verificación OK

* Cuadro Dieta OK

* Cuadro de diálogo y atributo sexo añadidos

* Lectura y escritura de ficheros JSON añadida. Generador de PDF en progreso. Cuadro de dialogo para confirmar decisiones añadido.

* Fichero_rutina.json añadido.

---------

+++ src/View/Menu_principal.java
@@ -1,13 +1,14 @@
 package View;
 
 import Controller.Controller;
-
+import Model.Gestion_de_ficheros;
 import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
 public class Menu_principal {
 
+    static Gestion_de_ficheros gestionDeFicheros = new Gestion_de_ficheros();
     static Controller controlador = new Controller();
     private static JFrame ventana = new JFrame("Menu principal");
     private JPanel panel1;
@@ -33,6 +34,11 @@ public class Menu_principal {
         //Frase personalizada para el usuario en el header del menu principal//
         String texto_etiqueta = "Bienvenid@ " + controlador.getCliente_sesion_actual().getUsername() + ". ¿Que deseas hacer?";
         etiqueta_custom.setText(texto_etiqueta);
+
+        if (gestionDeFicheros.comprobar_existencia_dieta_rutina("Fichero_dieta.json")==false || gestionDeFicheros.comprobar_existencia_dieta_rutina("Fichero_rutina.json")==false){
+            verPlanningActualButton.setEnabled(false);
+            descargarPlanningButton.setEnabled(false);
+        }
         nuevaDietaButton.addActionListener(new ActionListener() {
             /**
              * @param e the event to be processed
@@ -49,7 +55,15 @@ public class Menu_principal {
              */
             @Override
             public void actionPerformed(ActionEvent e) {
-
+                Cuadro_de_dialogo dialogo = new Cuadro_de_dialogo();
+                boolean aux = dialogo.confirmacion_de_decision_2("Se generará un nuevo entrenamiento con los siguientes datos: \n Objetivo " +
+                        "(" + controlador.getCliente_sesion_actual().getTipo_dieta() + ") , Experiencia en el gimansio (" + controlador.getCliente_sesion_actual().getExperiencia() + ") " +
+                        " Peso (" + controlador.getCliente_sesion_actual().getPeso() + "). ¿Estas de acuerdo?" );
+                if (aux==true){
+                    //añadir metodo de nueva rutina de ejercicios//
+                    controlador.escribir_ficheros(new String[][]{{"hola", "chao", "biseps"}, {"hola", "chao", "biseps"}}, "Fichero_rutina.json");
+                    JOptionPane.showMessageDialog(null, "Se ha generado una nueva rutina de ejercicios." );
+                }
             }
         });
         verPlanningActualButton.addActionListener(new ActionListener() {
@@ -58,7 +72,8 @@ public class Menu_principal {
              */
             @Override
             public void actionPerformed(ActionEvent e) {
-
+                Planning.lanzar_ventana();
+               ventana.dispose();
             }
         });
         descargarPlanningButton.addActionListener(new ActionListener() {
@@ -67,6 +82,7 @@ public class Menu_principal {
              */
             @Override
             public void actionPerformed(ActionEvent e) {
+                controlador.generar_pdf();
 
             }
         });
@@ -95,10 +111,16 @@ public class Menu_principal {
              */
             @Override
             public void actionPerformed(ActionEvent e) {
-                controlador.eliminar_cliente(controlador.getCliente_sesion_actual().getIdCliente());
-                JOptionPane.showMessageDialog(null, "Tu perfil ha sido eliminado. Volverás a la pantalla de Inicio");
-                Inicio.lanzar_ventana();
-                ventana.dispose();
+                Cuadro_de_dialogo dialogo = new Cuadro_de_dialogo();
+                if (dialogo.confirmacion_de_decision("¿Seguro que deseas eliminar tu perfil?")){
+                    controlador.eliminar_cliente(controlador.getCliente_sesion_actual().getIdCliente());
+                    JOptionPane.showMessageDialog(null, "Tu perfil ha sido eliminado. Volverás a la pantalla de Inicio");
+                    Inicio.lanzar_ventana();
+                    ventana.dispose();
+                } else {
+                    System.out.println("No se ha eliminado el perfil");
+                }
+
             }
         });
         cerrarSesiónButton.addActionListener(new ActionListener() {
@@ -114,7 +136,8 @@ public class Menu_principal {
     }
 
     /**
-     * Lanza la ventana de la interfaz grafica con una configuracion determinada e iniciando el constructor de la clase, lo que activa los listeners
+     * Lanza la ventana de la interfaz gráfica e inicializa el constructor de la clase
+     * inicia los listener de esta interfaz gráfica
      */
     public static void lanzar_ventana(){
         ventana.setContentPane(new Menu_principal().panel1);
